+++ crates/dioxus-daisyui/src/colors.rs
@@ -1,0 +1,20 @@
+use crate::ext::*;
+
+// https://daisyui.com/docs/colors/
+semantic_colors!(bg);
+semantic_colors!(to);
+semantic_colors!(via);
+semantic_colors!(from);
+semantic_colors!(text);
+semantic_colors!(ring);
+semantic_colors!(fill);
+semantic_colors!(caret);
+semantic_colors!(stroke);
+semantic_colors!(border);
+semantic_colors!(divide);
+semantic_colors!(accent);
+semantic_colors!(shadow);
+semantic_colors!(outline);
+semantic_colors!(decoration);
+semantic_colors!(placeholder);
+semantic_colors!(ring offest);