+++ crates/dioxus-tailwindcss/src/accessibility.rs
@@ -1,0 +1,5 @@
+use crate::ext::*;
+
+// https://tailwindcss.com/docs/screen-readers
+constant!(sr only);
+constant!(not sr only);