+++ crates/dioxus-tailwindcss/src/effects.rs
@@ -14,23 +14,7 @@ constant!(shadow none);
 colors!(shadow);
 
 // https://tailwindcss.com/docs/opacity
-constant!(opacity 0);
-constant!(opacity 5);
-constant!(opacity 10);
-constant!(opacity 20);
-constant!(opacity 25);
-constant!(opacity 30);
-constant!(opacity 40);
-constant!(opacity 50);
-constant!(opacity 60);
-constant!(opacity 70);
-constant!(opacity 75);
-constant!(opacity 80);
-constant!(opacity 90);
-constant!(opacity 95);
-constant!(opacity 100);
-
-any!(opacity);
+opacities!(opacity);
 
 // https://tailwindcss.com/docs/mix-blend-mode
 blend!(mix blend);