+++ crates/dioxus-tailwindcss/src/filters.rs
@@ -1,0 +1,163 @@
+use crate::ext::*;
+
+// https://tailwindcss.com/docs/blur
+constant!(blur none);
+constant!(blur);
+sm_to_2xl!(blur);
+constant!(blur 3xl);
+
+any!(blur);
+
+// https://tailwindcss.com/docs/brightness
+constant!(brightness 0);
+constant!(brightness 50);
+constant!(brightness 75);
+constant!(brightness 90);
+constant!(brightness 95);
+constant!(brightness 100);
+constant!(brightness 105);
+constant!(brightness 110);
+constant!(brightness 125);
+constant!(brightness 150);
+constant!(brightness 200);
+
+any!(brightness);
+// https://tailwindcss.com/docs/contrast
+constant!(contrast 0);
+constant!(contrast 50);
+constant!(contrast 75);
+constant!(contrast 100);
+constant!(contrast 110);
+constant!(contrast 125);
+constant!(contrast 150);
+constant!(contrast 200);
+
+any!(contrast);
+
+// https://tailwindcss.com/docs/drop-shadow
+constant!(drop shadow);
+sm_to_2xl!(drop shadow);
+constant!(drop shadow none);
+
+any!(drop shadow);
+
+// https://tailwindcss.com/docs/grayscale
+constant!(grayscale);
+constant!(grayscale 0);
+
+any!(grayscale);
+
+// https://tailwindcss.com/docs/hue-rotate
+constant!(hue rotate 0);
+constant!(hue rotate 15);
+constant!(hue rotate 30);
+constant!(hue rotate 60);
+constant!(hue rotate 90);
+constant!(hue rotate 180);
+
+any!(hue rotate);
+
+// https://tailwindcss.com/docs/invert
+constant!(invert);
+constant!(invert 0);
+
+any!(invert);
+
+// https://tailwindcss.com/docs/saturate
+constant!(saturate);
+constant!(saturate 0);
+constant!(saturate 50);
+constant!(saturate 100);
+constant!(saturate 150);
+constant!(saturate 200);
+
+any!(saturate);
+
+// https://tailwindcss.com/docs/sepia
+constant!(sepia);
+constant!(sepia 0);
+
+any!(sepia);
+
+// https://tailwindcss.com/docs/backdrop-blur
+constant!(backdrop blur none);
+constant!(backdrop blur);
+sm_to_2xl!(backdrop blur);
+constant!(backdrop blur 3xl);
+
+any!(backdrop blur);
+
+// https://tailwindcss.com/docs/backdrop-brightness
+constant!(backdrop brightness 0);
+constant!(backdrop brightness 50);
+constant!(backdrop brightness 75);
+constant!(backdrop brightness 90);
+constant!(backdrop brightness 95);
+constant!(backdrop brightness 100);
+constant!(backdrop brightness 105);
+constant!(backdrop brightness 110);
+constant!(backdrop brightness 125);
+constant!(backdrop brightness 150);
+constant!(backdrop brightness 200);
+
+any!(backdrop brightness);
+// https://tailwindcss.com/docs/backdrop-contrast
+constant!(backdrop contrast 0);
+constant!(backdrop contrast 50);
+constant!(backdrop contrast 75);
+constant!(backdrop contrast 100);
+constant!(backdrop contrast 110);
+constant!(backdrop contrast 125);
+constant!(backdrop contrast 150);
+constant!(backdrop contrast 200);
+
+any!(backdrop contrast);
+
+// https://tailwindcss.com/docs/backdrop-drop-shadow
+constant!(backdrop drop shadow);
+sm_to_2xl!(backdrop drop shadow);
+constant!(backdrop drop shadow none);
+
+any!(backdrop drop shadow);
+
+// https://tailwindcss.com/docs/backdrop-grayscale
+constant!(backdrop grayscale);
+constant!(backdrop grayscale 0);
+
+any!(backdrop grayscale);
+
+// https://tailwindcss.com/docs/backdrop-hue-rotate
+constant!(backdrop hue rotate 0);
+constant!(backdrop hue rotate 15);
+constant!(backdrop hue rotate 30);
+constant!(backdrop hue rotate 60);
+constant!(backdrop hue rotate 90);
+constant!(backdrop hue rotate 180);
+
+any!(backdrop hue rotate);
+
+// https://tailwindcss.com/docs/backdrop-invert
+constant!(backdrop invert);
+constant!(backdrop invert 0);
+
+any!(backdrop invert);
+
+// https://tailwindcss.com/docs/backdrop-opacity
+opacities!(backdrop opacity);
+
+// https://tailwindcss.com/docs/backdrop-saturate
+constant!(backdrop saturate);
+constant!(backdrop saturate 0);
+constant!(backdrop saturate 50);
+constant!(backdrop saturate 100);
+constant!(backdrop saturate 150);
+constant!(backdrop saturate 200);
+
+any!(backdrop saturate);
+
+// https://tailwindcss.com/docs/backdrop-sepia
+constant!(backdrop sepia);
+constant!(backdrop sepia 0);
+
+any!(backdrop sepia);
+