+++ crates/dioxus-tailwindcss/src/svg.rs
@@ -1,0 +1,14 @@
+use crate::ext::*;
+
+// https://tailwindcss.com/docs/fill
+constant!(fill none);
+colors!(fill);
+
+// https://tailwindcss.com/docs/stroke
+constant!(stroke none);
+colors!(stroke);
+
+// https://tailwindcss.com/docs/stroke-width
+constant!(stroke 0);
+constant!(stroke 1);
+constant!(stroke 2);